requirements.txt
telethon
discord.py
peewee
pycryptodome
discord-py-slash-command